[Refactor] Board 페이지 컴포넌트화 커밋
