+++ src/components/board/BoardMainCategory.style.js
@@ -1,0 +1,33 @@
+import styled from "styled-components";
+
+const Around = styled.div`
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+`;
+
+const MainCategory = styled.div`
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  padding-bottom: 0.75rem;
+  text-decoration: none;
+
+  color: ${(props) =>
+    props.isActive ? "rgb(70, 70, 70)" : "rgb(180, 180, 180)"};
+  border-bottom: ${(props) =>
+    props.isActive
+      ? "0.2rem solid rgb(70, 70, 70)"
+      : "0.1rem solid rgb(180, 180, 180)"};
+
+  &:hover {
+    cursor: pointer;
+  }
+
+  &:active {
+    color: rgb(70, 70, 70);
+    border-bottom: 0.2rem solid rgb(70, 70, 70);
+  }
+`;
+
+export { Around, MainCategory }
