+++ src/components/elem/Icons.jsx
@@ -6,7 +6,7 @@ const Icons = [
   "🎨",
   "⚽️",
   "✈️",
-  "👩‍❤️‍👨",
+  "💍",
   "🎮",
   "🐶",
   "👨‍👩‍👧‍👦",
