+++ src/pages/board/TotalBoard.jsx
@@ -1,12 +1,15 @@
-import React, { useState, useRef } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import Footer from "../../components/footer/Footer";
 import styled from "styled-components";
-import BoardHeader from "../../components/header/BoardHeader";
 import MobileLayout from "../../layout/MobileLayout";
 import BoardPostCard from "../../components/post/BoardPostCard";
 import InfiniteScroll from "../../components/hook/scroll/InfiniteScroll";
-import TotalBoardFetch from "../../components/hook/fetch/TotalBoardFetch";
+import BoardMainCategory from "../../components/board/BoardMainCategory";
+import BoardCategorySlider from "../../components/board/BoardCategorySlider";
 import { Helmet } from "react-helmet";
+import { instance } from "../../api/axios";
+import { useLocation } from "react-router-dom";
+import * as St from "../../components/board/Board.style";
 
 const TotalBoard = () => {
   const BoardCallback = (x, y, z) => {
@@ -14,13 +17,50 @@ const TotalBoard = () => {
     setCategory(y);
     setPage(z);
   };
-  const postCardContRef = useRef(null);
+
+  // const MainCatCallback = (x) => {
+  //   set
+  // }
+
   const [prevCategory, setPrevCategory] = useState("전체");
   const [category, setCategory] = useState("전체");
+  // const [maincategory, setMaincategory] = useState("전체");
   const [page, setPage] = useState(1);
   const [data, setData] = useState([]);
+  const location = useLocation();
 
-  TotalBoardFetch({ prevCategory, category, setData, page });
+  useEffect(() => {
+    const fetchData = async () => {
+      const maincategory =
+        location.pathname === "/totalboard"
+          ? "전체"
+          : location.pathname === "/humourboard"
+          ? "유머"
+          : "진지";
+
+      try {
+        const response = await instance.get(
+          `/postCards?maincategory=${maincategory}&category=${category}&splitNumber=7&splitPageNumber=${page}`
+        );
+        if (prevCategory === category) {
+          setData((prev) => [...prev, ...response.data.postCards]);
+          if (response.data.postCards.length === 0) {
+            alert("마지막 게시물입니다.");
+          }
+        } else {
+          setData(response.data.postCards);
+          if (response.data.postCards.length === 0) {
+            alert(`${category} 게시물이 없습니다.`);
+          }
+        }
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    fetchData();
+  }, [category, page]);
+
+  const postCardContRef = useRef(null);
   InfiniteScroll({ postCardContRef, setPage });
 
   return (
@@ -29,7 +69,14 @@ const TotalBoard = () => {
         <title>훈수 — 전체게시판</title>
       </Helmet>
       <MobileLayout>
-        <BoardHeader parentFunction={BoardCallback} />
+        <St.Header>
+          <St.Wrap>
+            <St.Main>훈수게시판</St.Main>
+            {/* <BoardMainCategory parentFunction={MainCatCallback}/> */}
+            <BoardMainCategory />
+            <BoardCategorySlider parentFunction={BoardCallback} />
+          </St.Wrap>
+        </St.Header>
         <PostCardList ref={postCardContRef}>
           {data?.map((item) => (
             <BoardPostCard
@@ -43,7 +90,6 @@ const TotalBoard = () => {
               commentCount={item.commentCount}
               mainCategory={item.maincategory}
               category={item.category}
-              // isFirst={idx === 0}
             />
           ))}
         </PostCardList>
